requirements.txt
PyYAML==3.12
appdirs==1.4.0
backports-abc==0.5
backports.ssl-match-hostname==3.5.0.1
certifi==2017.1.23
cffi==1.9.1
cryptography==1.7.2
enum34==1.1.6
idna==2.2
ipaddress==1.0.18
gnureadline==6.3.8
packaging==16.8
paramiko==2.1.1
pyasn1==0.2.2
pycparser==2.17
pyparsing==2.1.10
requests==2.13.0
singledispatch==3.4.0.3
six==1.10.0
urllib3==1.22
click==6.7